complete 1st assignment
It is quite possible something here is wrong and I have
misuderstood the instructions. For some reason they just do not
_click_. This whole assignment is done under assumption that what
we actually want to run Chandy-Lamport algorithm is a strongly
connected graph. This assumption was concluded from personal
research. The lectures and the instructions make it sound like that
is not the case but they are also very vague on what makes a good
initiator. Additionally this project uses petgraph crate to
represent the graph data structure and get the structure to be
serializable which might not be what the professor envisioned but it
was not forbidden either. The performance of the solution is optimized
with the second part of the assignment in mind.